# src/main/resources/static/api-docs.yml
swagger: "2.0"
info:
  description: >-
    Especificaciones técnicas para consumo de la api DigitalBanking
  version: 1.0.1
  title: Rest Api DigitalBanking
  termsOfService: http://swagger.io/terms/
  contact:
    email: soporte@digitalbanking.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:8080
basePath: /api/v1
...
schemes:
  - https
  - http
paths:
...
  /rol/buscarTodos:
    post:
      tags:
        - Roles
      summary: Obtiene todos las roles
      description: ""
      operationId: buscarRoles
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/PagineableDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            type: array
            items:
              $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
  /rol/{id}:
    get:
      tags:
        - Roles
      summary: Obtiene un rol por id
      description: ""
      operationId: buscarRolPorId
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          required: true
          name: id
          description: id
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    delete:
      tags:
        - Roles
      summary: Elimina un rol
      description: ""
      operationId: eliminarRol
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: id
          required: true
          description: id
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
  /usuario/buscarTodos:
    post:
      tags:
        - Usuarios
      summary: Busca todos los usuarios registrados en el sistema
      description: ""
      operationId: buscarUsuarios
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/PagineableDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            type: array
            items:
              $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /usuario/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtiene un usuario por id
      description: ""
      operationId: buscarUsuarioPorId
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    put:
      tags:
        - Usuarios
      summary: Actualiza un usuario
      description: ""
      operationId: actualizarUsuario
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/ActualizarUsuarioDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /usuario/correo/{correo}:
    get:
      tags:
        - Usuarios
      summary: Obtiene un usuario por correo
      description: ""
      operationId: buscarUsuarioPorCorreo
      produces:
        - application/json
      parameters:
        - in: path
          name: correo
          description: correo
          required: true
          type: string
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /usuario:
    post:
      tags:
        - Usuarios
      summary: Crear un usuario
      description: ""
      operationId: crearUsuario
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/RegistroRequestDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: Authorization
definitions:
  PagineableDto:
    type: object
    properties:
      pagina:
        type: integer
        format: int32
      cantidad:
        type: integer
        format: int32
...
  RolDto:
    type: object
    properties:
      idRol:
        type: integer
        format: int32
      codigo:
        type: string
      nombre:
        type: string
      indicadorHabilitado:
        type: boolean
  UsuarioDto:
    type: object
    properties:
      idUsuario:
        type: integer
      usuario:
        type: string
      rol:
        type: object
        properties:
          nombre:
            type: string
      persona:
        type: object
        properties:
          nombre:
            type: string
          apellido:
            type: string
          correo:
            type: string
      indicadorHabilitado:
        type: boolean
      verificado:
        type: boolean
      tokenValidacionCorreo:
        type: string
      fechaExpiracionToken:
        type: string
        format: datetime
  RegistroRequestDto:
    type: object
    properties:
      nombre:
        type: string
      apellido:
        type: string
      correo:
        type: string
      usuario:
        type: string
      contrasenia:
        type: string
      rol:
        type: integer
        format: int32
  ActualizarUsuarioDto:
    type: object
    properties:
      idUsuario:
        type: integer
      nombre:
        type: string
      apellido:
        type: string
      correo:
        type: string
      usuario:
        type: string
      contrasenia:
        type: string
      rol:
        type: integer
        format: int32
      indicadorHabilitado:
        type: boolean
      verificado:
        type: boolean
...
  Error:
    type: object
    properties:
      timestamp:
        type: integer
        format: int32
      status:
        type: integer
        format: int32
      error:
        type: string
      path:
        type: string
...